# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API 문서",
    "version": "1.0",
    "description": "Flask 애플리케이션의 API 문서입니다. 사용자 관리, 메시지, 스토어 관련 작업에 대한 엔드포인트를 포함합니다."
  },
  "host": "localhost:5001",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/register": {
      "post": {
        "summary": "사용자 등록",
        "description": "이름, 학번, 비밀번호 등의 정보를 입력하여 사용자를 등록합니다.",
        "tags": ["사용자"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "이예나"
                },
                "studentID": {
                  "type": "string",
                  "example": "1111111111"
                },
                "password": {
                  "type": "string",
                  "example": "1111"
                },
                "choiceType": {
                  "type": "string",
                  "example": "슈붕"
                },
                "topic": {
                  "type": "string",
                  "example": "토픽1 어쩌고 저쩌고"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "사용자 등록 성공"
          },
          "400": {
            "description": "잘못된 입력 또는 이미 존재하는 사용자"
          },
          "500": {
            "description": "회원 가입 오류"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "사용자 로그인",
        "description": "학번과 비밀번호를 입력하여 사용자가 로그인할 수 있습니다.",
        "tags": ["사용자"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "studentID": {
                  "type": "string",
                  "example": "1111111111"
                },
                "password": {
                  "type": "string",
                  "example": "1111"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "로그인 성공",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "로그인 성공"
                },
                "name": {
                  "type": "string",
                  "example": "이예나"
                },
                "choiceType": {
                  "type": "string",
                  "example": "슈붕"
                },
                "token": {
                  "type": "string",
                  "example": "jwt_token_example"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 학번 또는 비밀번호",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "비밀번호가 일치하지 않습니다"
                }
              }
            }
          }
        }
      }
    },
    "/store/{userID}": {
      "get": {
        "summary": "사용자 스토어 조회",
        "description": "특정 사용자의 스토어 정보를 조회합니다.",
        "tags": ["스토어"],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "integer",
            "example": 1
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT 토큰 (형식: Bearer <token>)"
          }
        ],
        "responses": {
          "200": {
            "description": "스토어 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "Store for userID 1"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "이예나"
                    },
                    "choiceType": {
                      "type": "string",
                      "example": "슈붕"
                    },
                    "coin": {
                      "type": "integer",
                      "example": 1500
                    },
                    "memoChoiceCount": {
                      "type": "object",
                      "properties": {
                        "팥붕": {
                          "type": "integer",
                          "example": 3
                        },
                        "슈붕": {
                          "type": "integer",
                          "example": 2
                        },
                        "민초붕": {
                          "type": "integer",
                          "example": 1
                        },
                        "고구마붕": {
                          "type": "integer",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자 없음",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User with ID 1 not found"
                }
              }
            }
          }
        }
      }
    },
    "/store/{userID}/write/{type}": {
      "post": {
        "summary": "스토어에 메시지 작성",
        "description": "다른 사용자의 스토어에 메시지를 작성합니다.",
        "tags": ["스토어"],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "integer",
            "example": 2
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "팥붕"
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT 토큰 (형식: Bearer <token>)"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "안녕하세요! 좋은 하루 되세요."
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "메시지 작성 성공"
          },
          "400": {
            "description": "본인의 가게에 메시지를 작성할 수 없음"
          },
          "404": {
            "description": "사용자 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/all-store": {
      "get": {
        "summary": "모든 스토어 조회",
        "description": "모든 사용자 목록과 정보를 가져옵니다.",
        "tags": ["스토어"],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT 토큰 (형식: Bearer <token>)"
          }
        ],
        "responses": {
          "200": {
            "description": "모든 사용자 목록 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "store_list": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userID": {
                            "type": "integer",
                            "example": 1
                          },
                          "username": {
                            "type": "string",
                            "example": "이예나"
                          },
                          "type": {
                            "type": "string",
                            "example": "슈붕"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자 없음"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "사용자",
      "description": "사용자 관리 관련 엔드포인트"
    },
    {
      "name": "스토어",
      "description": "스토어 관련 엔드포인트"
    }
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT 토큰 (형식: Bearer <token>)"
    }
  }
}